refactor(ui): move data-display primitives into shared/ui/data-display
