+++ src/app/home-page.tsx
@@ -9,7 +9,10 @@ import { ErrorMessage } from "@/shared/ui/error-message";
 import { Loading } from "@/shared/ui/feedback/loading";
 import { useAllPosts } from "@/features/post/model/use-all-posts";
 import { useHomeStats } from "@/features/post/model/use-home-stats";
-import { Sparkline, seriesFromMonths } from "@/shared/ui/sparkline";
+import {
+  Sparkline,
+  seriesFromMonths,
+} from "@/shared/ui/data-display/sparkline";
 import { Label, Category, Metric, StatusBadge, Dot } from "@/shared/ui";
 import { MatrixText } from "@/shared/ui/effects";
 import { useInfiniteScroll } from "@/shared/lib/use-infinite-scroll";
