+++ src/features/arcade/snake/ui/snake-metrics.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { Label, Spinner } from "@/shared/ui";
-import { Sparkline } from "@/shared/ui/sparkline";
+import { Sparkline } from "@/shared/ui/data-display/sparkline";
 import { formatScore } from "../model/leaderboard";
 import { HISTORY_RECENT } from "../model/score-history";
 import type { HistoryPoint, SnakeGameState } from "../model/types";
