+++ src/shared/ui/data-display/avatar.tsx
