+++ src/shared/ui/data-display/category.tsx
