+++ src/shared/ui/data-display/dot.tsx
