+++ src/shared/ui/data-display/index.ts
@@ -1,0 +1,5 @@
+export { Avatar } from "./avatar";
+export { Metric, fmt } from "./metric";
+export { Category } from "./category";
+export { Dot } from "./dot";
+export { StatBar } from "./stat-bar";
