+++ src/shared/ui/data-display/metric.tsx
