+++ src/shared/ui/data-display/sparkline.test.ts
