+++ src/shared/ui/data-display/sparkline.tsx
