+++ src/shared/ui/data-display/stat-bar.tsx
